// src/modules/auth/auth.schema.ts
import {email, z} from 'zod';
export const signupSchema = z.object({
    email: z.email({message: "Invalid email address"}),
    password: z.string().min(6, {message: "Password must be at least 6 characters long"})
})

export const loginSchema = z.object({
    email: z.email({message: "Invalid email address"}),
    password: z.string().min(1, {message: "Password is required"})
})